refactor join page naming -> apply page
